+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "com.example.sufcwmillionsapplication"
+        applicationId "uk.ac.kcl.sufcwmillionapplication"
         minSdkVersion 16
         targetSdkVersion 29
         versionCode 1
