+++ app/src/androidTest/java/uk/ac/kcl/sufcwmillionapplication/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.sufcwmillionsapplication;
+package uk.ac.kcl.sufcwmillionapplication;
 
 import android.content.Context;
 
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
     public void useAppContext() {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("com.example.sufcwmillionsapplication", appContext.getPackageName());
+        assertEquals("uk.ac.kcl.sufcwmillionapplication", appContext.getPackageName());
     }
 }
