+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">SufCwMillionsApplication</string>
+    <string name="app_name">SufCwMillionApplication</string>
 </resources>
